Refactor and split S3Bucket.url_for
The methods to call are now S3Bucket.make_url and
S3Bucket.make_url_authed. url_for will still work, but will show a
deprecation warning.

Previous code would intertwine the testing of the expiry calculation
with the testing of the actual URL generator; this made for awkward mock
methods such as S3Bucket._now. This is now split and tested separately.