Add nose-based testing
Note: this means the doctests tests aren't run, which is sort of fine.
They require an intelligent mock S3 interface, which would introduce
immense complexity. That, or the real deal--neither is very friendly to
testing.

There's still lots of things to cover with testing, including delete,
listdir, create_bucket and delete_bucket. In most need of tests is
listdir, because it's most complex.